requirements.txt
Pillow==10.3.0
python-telegram-bot==20.7
requests==2.31.0
